computervision/requirements.txt
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
Pillow>=9.0.0
scikit-learn>=1.0.0
tqdm>=4.60.0
matplotlib>=3.5.0
